+++ pkg/analyzer/lib/src/dart/element/member.dart
@@ -7,6 +7,7 @@ import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:analyzer/src/dart/ast/ast.dart';
 import 'package:analyzer/src/dart/element/display_string_builder.dart';
 import 'package:analyzer/src/dart/element/element.dart';
 import 'package:analyzer/src/dart/element/type.dart';
@@ -1255,14 +1256,8 @@ abstract class VariableMember extends Member
       super.baseElement as VariableElementImpl;
 
   @override
-  ConstantInitializer? get constantInitializer {
+  ExpressionImpl? get constantInitializer {
     return baseElement.constantInitializer;
-  }
-
-  @Deprecated('Use constantInitializer instead')
-  @override
-  ConstantInitializer? get constantInitializer2 {
-    return constantInitializer;
   }
 
   @override