+++ pkg/analyzer/lib/src/summary2/detach_nodes.dart
@@ -100,7 +100,6 @@ class _Visitor extends GeneralizingElementVisitor2<void> {
         fragment.typeInference = null;
       }
     }
-    element.constantInitializer;
     _detachConstVariable(element);
     super.visitPropertyInducingElement(element);
   }